// frontend/src/app/customers/types.tsx
export interface CustomerType {
  customer_id: number;
  customer_username: string;
  customer_uuid: string;
  customer_position: string;
  customer_full_name: string;
  customer_phone_number: string;
  customer_email_address: string;
  customer_created_by_employee_id: number;
  created_at: string;
}